refactor tree travel to be in separate file
